requirements.txt
pylint
pytest
requests
pymssql
python-dotenv